Tracing a set segments
Adapted the code so it will display n segments at a time.
There are n+1 meshes in the viewer. One for the main mesh and n for the each of the streamline segments. For each itteration we replace the last segment of each streamline with the new first one.

Todo: Give the streamlines a lifespan