Implemented mask color based on effort
Used the igl::jet function because this represented the nullschool example a lot better then parula.

Also increased the number of noodles spawned from 0.3 spawn (default value) to 1.5/degree. This way the mesh is nicely covered with the noodles for all degrees of vector fields (same amount of noodles for all N)