Update according final meeting
Viewer related action returned to main.

The streamline are now a single mesh, instead of a mesh per set of segments. updated all related function to be able to work with this new structure

Updated the names of the function.

Deleted minimal/unnecessary functions.

Coloring changed from jet to parula.

updated the tutorial to include 104 dynamic visualization.